third_party/WebKit/LayoutTests/accessibility/deleting-iframe-destroys-axcache-expected.txt

Before

After

End of test

This tests that deleting an iframe doesn't cause the accessibility cache to be destroyed and recreated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE

Before:
AXRole: AXWebArea AXValue: 
    AXRole: AXParagraph AXValue: 
        AXRole: AXStaticText AXValue: Before
            AXRole: AXInlineTextBox AXValue: Before
    AXRole: AXGroup AXValue: 
        AXRole: AXUnknown AXValue: 
            AXRole: AXScrollArea AXValue: 
                AXRole: AXWebArea AXValue: 
                    AXRole: AXGroup AXValue: 
                        AXRole: AXButton AXValue: 
    AXRole: AXParagraph AXValue: 
        AXRole: AXStaticText AXValue: After
            AXRole: AXInlineTextBox AXValue: After
    AXRole: AXParagraph AXValue: 
        AXRole: AXStaticText AXValue: End of test

After:
AXRole: AXWebArea AXValue: 
    AXRole: AXParagraph AXValue: 
        AXRole: AXStaticText AXValue: Before
            AXRole: AXInlineTextBox AXValue: Before
    AXRole: AXParagraph AXValue: 
        AXRole: AXStaticText AXValue: After
            AXRole: AXInlineTextBox AXValue: After
    AXRole: AXParagraph AXValue: 
        AXRole: AXStaticText AXValue: End of test

PASS frameBodyRole == frameBody.role is false
PASS frameGroupRole == frameGroup.role is false
PASS frameButtonRole == frameButton.role is false
PASS before.isEqual(newBefore) is true
PASS after.isEqual(newAfter) is true

TEST COMPLETE
